Locator - allow filter by LocationCategory
In CMS Settings tab on Locator page, added following:

DisplayCategory boolean
has_one LocationCategory

If a user checks the box and selects a category, list is filtered by that category on load, and category drop down is removed.

Allows module to have multiple Locator pages in a site and show different locations on each.